requirements.txt
fastapi==0.70.0
motor==2.5.1
uvicorn==0.15.0
websockets==10.0
python-dotenv==0.19.1
